+++ libnopegl/src/backends/gl/program_gl_utils.c
@@ -38,7 +38,7 @@ int ngli_program_gl_set_locations_and_bindings(struct program *s,
     const char *name = NULL;
     int need_relink = 0;
     const struct pipeline_layout layout = ngli_pgcraft_get_pipeline_layout(crafter);
-    for (int i = 0; i < layout.nb_attributes; i++) {
+    for (size_t i = 0; i < layout.nb_attributes; i++) {
         const struct pipeline_attribute_desc *attribute_desc = &layout.attributes_desc[i];
         if (name && !strcmp(name, attribute_desc->name))
             continue;
@@ -53,7 +53,7 @@ int ngli_program_gl_set_locations_and_bindings(struct program *s,
     if (need_relink)
         ngli_glLinkProgram(gl, s_priv->id);
 
-    for (int i = 0; i < layout.nb_buffers; i++) {
+    for (size_t i = 0; i < layout.nb_buffers; i++) {
         const struct pipeline_buffer_desc *buffer_desc = &layout.buffers_desc[i];
         if (buffer_desc->type != NGLI_TYPE_UNIFORM_BUFFER)
             continue;
